requirements.txt
aiofiles>=23.2.1
aiohttp>=3.8.4
aiohttp-socks>=0.8.4
aiosignal>=1.3.1
annotated-types>=0.6.0
anthropic>=0.19.2
anyio>=3.7.1
asgiref>=3.7.2
async-property>=0.2.2
async-timeout>=4.0.2
async-tio>=1.3.2
asyncstdlib>=3.10.9
attrs>=23.1.0
beautifulsoup4>=4.12.3
blinker>=1.7.0
bottle>=0.12.25
Brotli>=1.1.0
browser-cookie3>=0.19.1
cairocffi>=1.6.1
CairoSVG>=2.7.1
certifi>=2023.5.7
cffi>=1.15.1
charset-normalizer>=3.1.0
claude-api-py>=0.0.5
click>=8.1.7
clr-loader>=0.2.6
colorama>=0.4.6
cssselect2>=0.7.0
curl_cffi>=0.6.2
defusedxml>=0.7.1
distro>=1.9.0
duckduckgo_search>=5.1.0
exceptiongroup>=1.1.1
fastapi>=0.104.1
filelock>=3.13.1
Flask>=3.0.0
frozenlist>=1.3.3
fsspec>=2024.2.0
g4f>=0.2.7.1
h11>=0.14.0
h2>=4.1.0
hpack>=4.0.0
httpcore>=0.17.0
httpx>=0.24.1
huggingface-hub>=0.21.4
hyperframe>=6.0.1
idna>=3.4
iniconfig>=2.0.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
Js2Py>=0.74
libretranslatepy>=2.1.1
loguru>=0.7.2
lxml>=5.1.0
lz4>=4.3.2
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
mdurl>=0.1.2
mmh3>=4.0.1
multidict>=6.0.4
nest-asyncio>=1.5.8
numpy>=1.26.2
openai>=0.27.8
OpenAIAuth>=2.0.0
outcome>=1.3.0.post0
packaging>=23.1
Pillow>=10.1.0
platformdirs>=4.0.0
pluggy>=1.0.0
plyer>=2.1.0
prompt-toolkit>=3.0.38
proxy-tools>=0.1.0
py-arkose-generator>=0.0.0.2
pycparser>=2.21
pycryptodome>=3.19.0
pycryptodomex>=3.19.0
pydantic>=2.5.1
pydantic_core>=2.14.3
PyExecJS>=1.5.1
Pygments>=2.15.1
pyjsparser>=2.7.1
PySocks>=1.7.1
pytest>=7.3.1
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
python-socks>=2.4.4
pythonnet>=3.0.3
pywebview>=5.0.5
PyYAML>=6.0.1
regex>=2023.6.3
requests>=2.31.0
revChatGPT>=6.7.6
rich>=13.4.2
selenium>=4.15.2
six>=1.16.0
sniffio>=1.3.0
socksio>=1.0.0
sortedcontainers>=2.4.0
soupsieve>=2.5
sseclient-py>=1.8.0
starlette>=0.27.0
tiktoken>=0.4.0
tinycss2>=1.2.1
tokenizers>=0.15.2
tomli>=2.0.1
tqdm>=4.65.0
translate>=3.6.1
trio>=0.23.1
trio-websocket>=0.11.1
typing_extensions>=4.8.0
tzdata>=2023.3
tzlocal>=5.2
urllib3>=2.0.3
uvicorn>=0.24.0.post1
waitress>=2.1.2
wcwidth>=0.2.6
webencodings>=0.5.1
websockets>=12.0
Werkzeug>=3.0.1
win32-setctime>=1.1.0
wsproto>=1.2.0
yarl>=1.9.2